maitenance(renderer): fixed warnings + improved some readability/performance

+++ engine/src/renderer/SubTexture2D.cpp
@@ -25,8 +25,8 @@ namespace nexo::renderer {
 
     std::shared_ptr<SubTexture2D> createFromCoords(const std::shared_ptr<Texture2D> &texture, const glm::vec2 &coords, const glm::vec2 &cellSize, const glm::vec2 &spriteSize)
     {
-        glm::vec2 min = {(coords.x * cellSize.x) / texture->getWidth() , (coords.y * cellSize.y) / texture->getHeight()};
-        glm::vec2 max = {((coords.x + spriteSize.x) * cellSize.x) / texture->getWidth(), ((coords.y + spriteSize.y) * cellSize.y) / texture->getHeight()};
+        glm::vec2 min = {(coords.x * cellSize.x) / static_cast<float>(texture->getWidth()) , (coords.y * cellSize.y) / static_cast<float>(texture->getHeight())};
+        glm::vec2 max = {((coords.x + spriteSize.x) * cellSize.x) / static_cast<float>(texture->getWidth()), ((coords.y + spriteSize.y) * cellSize.y) / static_cast<float>(texture->getHeight())};
 
         return std::make_shared<SubTexture2D>(texture, min, max);
     }
